+++ src/Server/Authorization.php
@@ -1,5 +1,6 @@
 <?php namespace Dingo\OAuth2\Server;
 
+use Closure;
 use RuntimeException;
 use Dingo\OAuth2\Entity\Entity;
 use Dingo\OAuth2\ScopeValidator;
@@ -70,6 +71,13 @@ class Authorization {
 	 * @var array
 	 */
 	protected $responseTypes = [];
+
+	/**
+	 * Authorized callback used once access token is issued.
+	 * 
+	 * @var \Closure
+	 */
+	protected $authorizedCallback;
 
 	/**
 	 * Create a new Dingo\OAuth2\Server\Authorization instance.
@@ -165,7 +173,17 @@ class Authorization {
 		{
 			$refreshToken = $this->issueRefreshToken($accessToken);
 
-			$response['refresh_token'] = $refreshToken;
+			$response['refresh_token'] = $refreshToken->getToken();
+		}
+
+		// Fire the authorized callback that a developer can set. This is handy
+		// for developers that want to avoid prompting a user to authorize
+		// a client once they have already given the client permission.
+		if ($this->authorizedCallback instanceof Closure)
+		{
+			$client = $this->storage->get('client')->get($accessToken->getClientId());
+
+			call_user_func($this->authorizedCallback, $accessToken, $client);
 		}
 
 		// When making a response from an entity we may get some optional
@@ -190,7 +208,7 @@ class Authorization {
 
 		$expires = time() + $this->refreshTokenExpiration;
 
-		$this->storage->get('token')->create($refreshToken, 'refresh', $accessToken->getClientId(), $accessToken->getUserId(), $expires);
+		$refreshToken = $this->storage->get('token')->create($refreshToken, 'refresh', $accessToken->getClientId(), $accessToken->getUserId(), $expires);
 
 		$this->storage->get('token')->associateScopes($refreshToken, $accessToken->getScopes());
 
@@ -307,6 +325,19 @@ class Authorization {
 	public function getRequest()
 	{
 		return $this->request;
+	}
+
+	/**
+	 * Set the authorized callback.
+	 * 
+	 * @param  \Closure  $callback
+	 * @return \Dingo\OAuth2\Server\Authorization
+	 */
+	public function setAuthorizedCallback(Closure $callback)
+	{
+		$this->authorizedCallback = $callback;
+
+		return $this;
 	}
 
 	/**